Refactor file organisation
