--- .flutter-plugins 5
@@ -1,29 +1,0 @@
-# This is a generated file; do not edit or check into version control.
-cloud_firestore=/Users/brendan-macmini/.pub-cache/hosted/pub.dev/cloud_firestore-4.4.2/
-cloud_firestore_web=/Users/brendan-macmini/.pub-cache/hosted/pub.dev/cloud_firestore_web-3.3.2/
-firebase_auth=/Users/brendan-macmini/.pub-cache/hosted/pub.dev/firebase_auth-4.2.8/
-firebase_auth_web=/Users/brendan-macmini/.pub-cache/hosted/pub.dev/firebase_auth_web-5.2.7/
-firebase_core=/Users/brendan-macmini/.pub-cache/hosted/pub.dev/firebase_core-2.6.1/
-firebase_core_web=/Users/brendan-macmini/.pub-cache/hosted/pub.dev/firebase_core_web-2.2.1/
-firebase_database=/Users/brendan-macmini/.pub-cache/hosted/pub.dev/firebase_database-10.0.13/
-firebase_database_web=/Users/brendan-macmini/.pub-cache/hosted/pub.dev/firebase_database_web-0.2.1+22/
-path_provider=/Users/brendan-macmini/.pub-cache/hosted/pub.dev/path_provider-2.0.12/
-path_provider_android=/Users/brendan-macmini/.pub-cache/hosted/pub.dev/path_provider_android-2.0.22/
-path_provider_foundation=/Users/brendan-macmini/.pub-cache/hosted/pub.dev/path_provider_foundation-2.1.1/
-path_provider_linux=/Users/brendan-macmini/.pub-cache/hosted/pub.dev/path_provider_linux-2.1.8/
-path_provider_windows=/Users/brendan-macmini/.pub-cache/hosted/pub.dev/path_provider_windows-2.1.3/
-shared_preferences=/Users/brendan-macmini/.pub-cache/hosted/pub.dev/shared_preferences-2.0.15/
-shared_preferences_android=/Users/brendan-macmini/.pub-cache/hosted/pub.dev/shared_preferences_android-2.0.15/
-shared_preferences_ios=/Users/brendan-macmini/.pub-cache/hosted/pub.dev/shared_preferences_ios-2.1.1/
-shared_preferences_linux=/Users/brendan-macmini/.pub-cache/hosted/pub.dev/shared_preferences_linux-2.1.3/
-shared_preferences_macos=/Users/brendan-macmini/.pub-cache/hosted/pub.dev/shared_preferences_macos-2.0.5/
-shared_preferences_web=/Users/brendan-macmini/.pub-cache/hosted/pub.dev/shared_preferences_web-2.0.4/
-shared_preferences_windows=/Users/brendan-macmini/.pub-cache/hosted/pub.dev/shared_preferences_windows-2.1.3/
-sqflite=/Users/brendan-macmini/.pub-cache/hosted/pub.dev/sqflite-2.2.0+2/
-url_launcher=/Users/brendan-macmini/.pub-cache/hosted/pub.dev/url_launcher-6.1.5/
-url_launcher_android=/Users/brendan-macmini/.pub-cache/hosted/pub.dev/url_launcher_android-6.0.23/
-url_launcher_ios=/Users/brendan-macmini/.pub-cache/hosted/pub.dev/url_launcher_ios-6.1.0/
-url_launcher_linux=/Users/brendan-macmini/.pub-cache/hosted/pub.dev/url_launcher_linux-3.0.2/
-url_launcher_macos=/Users/brendan-macmini/.pub-cache/hosted/pub.dev/url_launcher_macos-3.0.2/
-url_launcher_web=/Users/brendan-macmini/.pub-cache/hosted/pub.dev/url_launcher_web-2.0.14/
-url_launcher_windows=/Users/brendan-macmini/.pub-cache/hosted/pub.dev/url_launcher_windows-3.0.3/
