--- .flutter-plugins-dependencies 5
@@ -1,1 +1,0 @@
-{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"cloud_firestore","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/cloud_firestore-4.4.2/","native_build":true,"dependencies":["firebase_core"]},{"name":"firebase_auth","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/firebase_auth-4.2.8/","native_build":true,"dependencies":["firebase_core"]},{"name":"firebase_core","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/firebase_core-2.6.1/","native_build":true,"dependencies":[]},{"name":"firebase_database","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/firebase_database-10.0.13/","native_build":true,"dependencies":["firebase_core"]},{"name":"path_provider_foundation","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/path_provider_foundation-2.1.1/","native_build":true,"dependencies":[]},{"name":"shared_preferences_ios","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/shared_preferences_ios-2.1.1/","native_build":true,"dependencies":[]},{"name":"sqflite","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/sqflite-2.2.0+2/","native_build":true,"dependencies":[]},{"name":"url_launcher_ios","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/url_launcher_ios-6.1.0/","native_build":true,"dependencies":[]}],"android":[{"name":"cloud_firestore","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/cloud_firestore-4.4.2/","native_build":true,"dependencies":["firebase_core"]},{"name":"firebase_auth","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/firebase_auth-4.2.8/","native_build":true,"dependencies":["firebase_core"]},{"name":"firebase_core","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/firebase_core-2.6.1/","native_build":true,"dependencies":[]},{"name":"firebase_database","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/firebase_database-10.0.13/","native_build":true,"dependencies":["firebase_core"]},{"name":"path_provider_android","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/path_provider_android-2.0.22/","native_build":true,"dependencies":[]},{"name":"shared_preferences_android","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/shared_preferences_android-2.0.15/","native_build":true,"dependencies":[]},{"name":"sqflite","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/sqflite-2.2.0+2/","native_build":true,"dependencies":[]},{"name":"url_launcher_android","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/url_launcher_android-6.0.23/","native_build":true,"dependencies":[]}],"macos":[{"name":"cloud_firestore","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/cloud_firestore-4.4.2/","native_build":true,"dependencies":["firebase_core"]},{"name":"firebase_auth","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/firebase_auth-4.2.8/","native_build":true,"dependencies":["firebase_core"]},{"name":"firebase_core","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/firebase_core-2.6.1/","native_build":true,"dependencies":[]},{"name":"firebase_database","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/firebase_database-10.0.13/","native_build":true,"dependencies":["firebase_core"]},{"name":"path_provider_foundation","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/path_provider_foundation-2.1.1/","native_build":true,"dependencies":[]},{"name":"shared_preferences_macos","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/shared_preferences_macos-2.0.5/","native_build":true,"dependencies":[]},{"name":"sqflite","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/sqflite-2.2.0+2/","native_build":true,"dependencies":[]},{"name":"url_launcher_macos","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/url_launcher_macos-3.0.2/","native_build":true,"dependencies":[]}],"linux":[{"name":"path_provider_linux","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/path_provider_linux-2.1.8/","native_build":false,"dependencies":[]},{"name":"shared_preferences_linux","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/shared_preferences_linux-2.1.3/","native_build":false,"dependencies":["path_provider_linux"]},{"name":"url_launcher_linux","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/url_launcher_linux-3.0.2/","native_build":true,"dependencies":[]}],"windows":[{"name":"path_provider_windows","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/path_provider_windows-2.1.3/","native_build":false,"dependencies":[]},{"name":"shared_preferences_windows","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/shared_preferences_windows-2.1.3/","native_build":false,"dependencies":["path_provider_windows"]},{"name":"url_launcher_windows","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/url_launcher_windows-3.0.3/","native_build":true,"dependencies":[]}],"web":[{"name":"cloud_firestore_web","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/cloud_firestore_web-3.3.2/","dependencies":["firebase_core_web"]},{"name":"firebase_auth_web","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/firebase_auth_web-5.2.7/","dependencies":["firebase_core_web"]},{"name":"firebase_core_web","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/firebase_core_web-2.2.1/","dependencies":[]},{"name":"firebase_database_web","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/firebase_database_web-0.2.1+22/","dependencies":["firebase_core_web"]},{"name":"shared_preferences_web","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/shared_preferences_web-2.0.4/","dependencies":[]},{"name":"url_launcher_web","path":"/Users/brendan-macmini/.pub-cache/hosted/pub.dev/url_launcher_web-2.0.14/","dependencies":[]}]},"dependencyGraph":[{"name":"cloud_firestore","dependencies":["cloud_firestore_web","firebase_core"]},{"name":"cloud_firestore_web","dependencies":["firebase_core","firebase_core_web"]},{"name":"firebase_auth","dependencies":["firebase_auth_web","firebase_core"]},{"name":"firebase_auth_web","dependencies":["firebase_core","firebase_core_web"]},{"name":"firebase_core","dependencies":["firebase_core_web"]},{"name":"firebase_core_web","dependencies":[]},{"name":"firebase_database","dependencies":["firebase_core","firebase_database_web"]},{"name":"firebase_database_web","dependencies":["firebase_core","firebase_core_web"]},{"name":"path_provider","dependencies":["path_provider_android","path_provider_foundation","path_provider_linux","path_provider_windows"]},{"name":"path_provider_android","dependencies":[]},{"name":"path_provider_foundation","dependencies":[]},{"name":"path_provider_linux","dependencies":[]},{"name":"path_provider_windows","dependencies":[]},{"name":"shared_preferences","dependencies":["shared_preferences_android","shared_preferences_ios","shared_preferences_linux","shared_preferences_macos","shared_preferences_web","shared_preferences_windows"]},{"name":"shared_preferences_android","dependencies":[]},{"name":"shared_preferences_ios","dependencies":[]},{"name":"shared_preferences_linux","dependencies":["path_provider_linux"]},{"name":"shared_preferences_macos","dependencies":[]},{"name":"shared_preferences_web","dependencies":[]},{"name":"shared_preferences_windows","dependencies":["path_provider_windows"]},{"name":"sqflite","dependencies":[]},{"name":"url_launcher","dependencies":["url_launcher_android","url_launcher_ios","url_launcher_linux","url_launcher_macos","url_launcher_web","url_launcher_windows"]},{"name":"url_launcher_android","dependencies":[]},{"name":"url_launcher_ios","dependencies":[]},{"name":"url_launcher_linux","dependencies":[]},{"name":"url_launcher_macos","dependencies":[]},{"name":"url_launcher_web","dependencies":[]},{"name":"url_launcher_windows","dependencies":[]}],"date_created":"2023-02-20 18:46:33.734684","version":"3.7.1"}
