+++ lib/pages/admin/admin_side_menu.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:seg_coursework_app/pages/admin/admin_choice_boards.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import '../visual_timetable/visual_timetable_interface.dart';
-import 'package:seg_coursework_app/pages/child_menu.dart';
+import 'package:seg_coursework_app/pages/child_menu/customizable_column.dart';
 
 /// The side-menu of the admin's UI
 class AdminSideMenu extends StatelessWidget {
